# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RecoveryOS API",
    "description": "Auditable subscription-recovery orchestration.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/voice/contacts": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "Start Voice Contact",
        "operationId": "start_voice_contact_v1_voice_contacts_post",
        "parameters": [
          {
            "name": "X-Recovery-Operator-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Recovery-Operator-Token"
            }
          },
          {
            "name": "X-RecoveryOS-Operator-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Recoveryos-Operator-Token"
            }
          },
          {
            "name": "X-RecoveryOS-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Recoveryos-Csrf-Token"
            }
          },
          {
            "name": "recoveryos_operator_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Recoveryos Operator Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartVoiceContactRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartVoiceContactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/voice/cases/{case_id}/contact-setup": {
      "put": {
        "tags": [
          "voice"
        ],
        "summary": "Configure Voice Contact",
        "operationId": "configure_voice_contact_v1_voice_cases__case_id__contact_setup_put",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "X-Recovery-Operator-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Recovery-Operator-Token"
            }
          },
          {
            "name": "X-RecoveryOS-Operator-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Recoveryos-Operator-Token"
            }
          },
          {
            "name": "X-RecoveryOS-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Recoveryos-Csrf-Token"
            }
          },
          {
            "name": "recoveryos_operator_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Recoveryos Operator Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoiceContactSetupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceEligibilityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/voice/cases/{case_id}/eligibility": {
      "get": {
        "tags": [
          "voice"
        ],
        "summary": "Voice Eligibility",
        "operationId": "voice_eligibility_v1_voice_cases__case_id__eligibility_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Case Id"
            }
          },
          {
            "name": "X-Recovery-Operator-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Recovery-Operator-Token"
            }
          },
          {
            "name": "X-RecoveryOS-Operator-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Recoveryos-Operator-Token"
            }
          },
          {
            "name": "X-RecoveryOS-CSRF-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Recoveryos-Csrf-Token"
            }
          },
          {
            "name": "recoveryos_operator_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Recoveryos Operator Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceEligibilityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/voice/twiml/{attempt_id}": {
      "post": {
        "tags": [
          "voice"
        ],
        "summary": "Twiml For Elevenlabs",
        "operationId": "twiml_for_elevenlabs_v1_voice_twiml__attempt_id__post",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Attempt Id"
            }
          },
          {
            "name": "x-twilio-signature",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Twilio-Signature"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StartVoiceContactRequest": {
        "properties": {
          "case_id": {
            "type": "string",
            "title": "Case Id"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 200,
            "minLength": 8,
            "title": "Idempotency Key"
          },
          "max_duration_seconds": {
            "type": "integer",
            "maximum": 180.0,
            "exclusiveMinimum": 0.0,
            "title": "Max Duration Seconds",
            "default": 180
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "case_id",
          "idempotency_key"
        ],
        "title": "StartVoiceContactRequest"
      },
      "StartVoiceContactResponse": {
        "properties": {
          "attempt_id": {
            "type": "string",
            "title": "Attempt Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "status": {
            "type": "string",
            "enum": [
              "SUBMITTED",
              "REJECTED",
              "UNCERTAIN"
            ],
            "title": "Status"
          },
          "reason_code": {
            "type": "string",
            "title": "Reason Code"
          },
          "provider_call_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Call Id"
          },
          "retry_permitted": {
            "type": "boolean",
            "const": false,
            "title": "Retry Permitted",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "attempt_id",
          "provider",
          "status",
          "reason_code"
        ],
        "title": "StartVoiceContactResponse"
      },
...
    }
  }
}